backend/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose
sqlalchemy
psycopg2-binary
groq
sentence-transformers
pydantic